requirements.txt
opencv_python==4.1.1.26
tornado==6.0.3
Keras==2.2.4
baidu_aip==2.2.18.0
tensorflow==1.14.0
matplotlib==3.1.1
numpy==1.17.3
requests==2.22.0
scikit_image==0.16.2
pypinyin==0.36.0
Pillow==6.2.1
skimage==0.0
